+++ test-suite/testkit/tests/inflating_currency/main.rs
@@ -20,7 +20,7 @@ use exonum::{
     crypto::{KeyPair, PublicKey},
     helpers::Height,
 };
-use exonum_testkit::{ApiKind, TestKit, TestKitApi, TestKitBuilder};
+use exonum_testkit::{ApiKind, Spec, TestKit, TestKitApi, TestKitBuilder};
 use pretty_assertions::assert_eq;
 use rand::Rng;
 
@@ -33,7 +33,7 @@ mod inflating_cryptocurrency;
 fn init_testkit() -> TestKit {
     TestKitBuilder::validator()
         .with_validators(4)
-        .with_default_rust_service(CurrencyService)
+        .with(Spec::new(CurrencyService).with_default_instance())
         .build()
 }